schema: move well-known types package
