+++ voc/schema/schema.go
@@ -39,4 +39,9 @@ const (
 	Time = Prefix + `Time`
 	// A combination of date and time of day in the form [-]CCYY-MM-DDThh:mm:ss[Z|(+|-)hh:mm] (see Chapter 5.4 of ISO 8601).
 	DateTime = Prefix + `DateTime`
+
+	// A class, also often called a 'Type'; equivalent to rdfs:Class.
+	Class = Prefix + "Class"
+	// A property, used to indicate attributes and relationships of some Thing; equivalent to rdf:Property.
+	Property = Prefix + "Property"
 )
